requirements.txt
Django>=5.0.6,<6
django-autocomplete-light>=3.11.0,<4
django-crispy-forms>=2.1,<3
django-extensions>=3.2.3,<4
django-filter>=24.2,<25
django-tables2>=2.7.0,<3
djangorestframework>=3.15.1,<4
django-guardian==2.4.0
mysqlclient==2.2.4
requests>=2.32.3,<3
